refactor: ship flat and hierarchical prototype variants only
The m2v engine scores each intent by cosine similarity between the query
embedding and the intent prototype, independent of how intents are
grouped. The parallel-argmax domain variant is therefore equivalent to
the flat engine.

Make HierarchicalPrototypeIntentStore standalone — it owns its own
per-domain sub-stores and domain-fingerprint router. Drop the domain
store class, the domain pipeline, its entry point, docs, and tests.
