+++ docs/index.md
@@ -44,7 +44,6 @@ The pipeline is designed as a **fallback or confidence-based matcher** — it is
 | [Configuration](configuration.md) | All configuration options |
 | [Pipeline Internals](pipeline.md) | Architecture and runtime behaviour |
 | [Prototype Strategies](strategies.md) | Scoring strategy reference for prototype mode |
-| [Domain Prototype Store](domain_store.md) | Domain-grouped parallel-argmax prototype matching |
 | [Hierarchical Prototype Store](hierarchical_store.md) | Two-stage domain-routed prototype matching |
 | [Models](models.md) | Available pre-trained models and benchmark results |
 | [Training](training.md) | How to gather data and train/retrain models |
